requirements.txt
certifi==2021.10.8
distlib==0.3.4
filelock==3.6.0
mysql-connector-python==8.0.28
pipenv==2022.1.8
platformdirs==2.5.1
python-decouple==3.6
six==1.16.0
virtualenv==20.13.2
virtualenv-clone==0.5.7
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
pytz==2021.3
yarl==1.7.2
